+++ hackacode/src/main/java/service/IEmpleadoService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface IEmpleadoService {
     List<Empleado> listAll();
     Empleado save(EmpleadoDto empleadoDto);
-    Empleado findById(Integer id);
+    Empleado findById(Long id);
     void delete(Empleado empleado);
-    boolean existsById(Integer id);
+    boolean existsById(Long id);
 }
